+++ tests/simulation.py
@@ -1,7 +1,7 @@
 """
 Tests for simulation.py
 """
-#pylint: disable=missing-function-docstring,missing-class-docstring,duplicate-code
+#pylint: disable=missing-function-docstring,missing-class-docstring,duplicate-code,unused-argument
 from unittest import TestCase
 from unittest.mock import MagicMock, patch
 
@@ -26,7 +26,6 @@ class TestSimulation(TestCase):
     def test_repr_deleted(self):
         self.model._deleted = True
         self.assertTrue('Deleted Object' in str(self.model))
-
 
     def test_create_service(self):
         res = self.model.create_service('test', 'intf', 22, foo='bar')
@@ -54,10 +53,20 @@ class TestSimulation(TestCase):
             self.model.control()
         self.assertTrue('requires action' in str(err.exception))
 
+    @patch('cumulus_air_sdk.air_sdk.simulation.Simulation.start')
+    def test_load(self, mock_start):
+        self.model.load()
+        mock_start.assert_called()
+
     @patch('cumulus_air_sdk.air_sdk.simulation.Simulation.control')
     def test_start(self, mock_control):
         self.model.start()
         mock_control.assert_called_with(action='load')
+
+    @patch('cumulus_air_sdk.air_sdk.simulation.Simulation.store')
+    def test_stop(self, mock_store):
+        self.model.stop()
+        mock_store.assert_called()
 
     @patch('cumulus_air_sdk.air_sdk.simulation.Simulation.control')
     def test_store(self, mock_control):
@@ -160,7 +169,8 @@ class TestSimulationApi(TestCase):
         self.assertEqual(res[1].id, 'xyz')
 
     @patch('cumulus_air_sdk.air_sdk.util.raise_if_invalid_response')
-    def test_create(self, mock_raise):
+    @patch('cumulus_air_sdk.air_sdk.util.validate_timestamps')
+    def test_create(self, mock_validate, mock_raise):
         self.client.post.return_value.json.return_value = {'id': 'abc'}
         res = self.api.create(topology='abc123')
         self.client.post.assert_called_with(f'{self.client.api_url}/simulation/',
@@ -168,6 +178,14 @@ class TestSimulationApi(TestCase):
         mock_raise.assert_called_with(self.client.post.return_value, status_code=201)
         self.assertIsInstance(res, simulation.Simulation)
         self.assertEqual(res.id, 'abc')
+        mock_validate.assert_called_with('Simulation created', expires_at=None, sleep_at=None)
+
+    @patch('cumulus_air_sdk.air_sdk.util.raise_if_invalid_response')
+    @patch('cumulus_air_sdk.air_sdk.util.validate_timestamps')
+    def test_create_timestamps(self, mock_validate, mock_raise):
+        self.api.create(topology='abc123', expires_at='expired', sleep_at='sleepy')
+        mock_validate.assert_called_with('Simulation created', expires_at='expired',
+                                         sleep_at='sleepy')
 
     def test_create_required_kwargs(self):
         with self.assertRaises(AttributeError) as err: