+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2021-02-26
+### Added
+- A warning will now be logged when simulations are created with `expires_at`/`sleep_at` dates in the past
+- `AirApi.Simulation.load()` (an alias for `AirApi.Simulation.start()`)
+- `AirApi.Simulation.stop()` (an alias for `AirApi.Simulation.store()`)
+### Fixed
+- `JSONDecodeError` raised when deleting simulation node instructions
+
 ## [2.0.4] - 2021-02-10
 ### Fixed
 - LazyLoading some objects caused unwanted `PATCH` calls